wrote a function that sets the value of a bit to 1 at a given index.

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -14,7 +14,7 @@ int set_bit(unsigned long int *n, unsigned int index)
 	{
 		return (-1);
 	}
-	mask = (unsigned long int) 1 << index;
+	mask = 1UL << index;
 	*n |= mask;
 	return (1);
 }
